split duckdb files for build constraints
